Update : Mainpage VAC pattern
